done dssp + giỏ hàng full

+++ src/main/java/com/example/demo/DAO/SanPhamRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.Model.SanPham;
 
-public interface SanPhamRepository extends JpaRepository<SanPham, Long> {
+public interface SanPhamRepository extends JpaRepository<SanPham, Integer> {
 
 }
